Restructure pdl rewrite implementation into own frontend
