+++ xdsl/frontend/pypdl/builder.py
@@ -1,0 +1,16 @@
+from dataclasses import dataclass
+
+from xdsl.dialects.builtin import ModuleOp
+from xdsl.frontend.pyast.builder import PyASTBuilder
+from xdsl.frontend.pypdl.transforms.func_to_pdl_rewrite import FuncToPdlRewrite
+
+
+@dataclass
+class PyPDLRewriteBuilder(PyASTBuilder):
+    """Builder for PDL rewrites from aspects of a Python function."""
+
+    def build(self) -> ModuleOp:
+        """Build a module from the builder state."""
+        module = super().build()
+        FuncToPdlRewrite().apply(ctx=None, op=module)
+        return module
